feat: align images in center when adding to slides
The images used to be scaled improperly when added to the slides and
were not aligned properly. This commit now fixes those issues.

However there are some sloppy math involved still. For example, to
get the top margin, I get the `top` attribute of the body text box.

This will be an issue in the future should I use a slide layout with no
body text placeholder