fix: solve edge case where no cropping is done
There was an issue where if a user did not interact with the cropping function,
there would be no cropped image to return to send to the server. This would
then crash the app.

How it is solved is by adding an initial cropping state that covers the entire picture,
so that even if the user does not crop, the entire image is captured.