feat(pptx): improve image aspect ratio maintaining function
Previously, the way the aspect ratio was maintained introduced unecessary
padding to the image which would show up as empty space if it was saved as an image.

So to fix that the image is now scaled manually instead of leaving it to
`pptxgenjs` library. It compares the scale value of both axis, and then
if one of the scale is smaller, it then sets that respective axis to fill up as much as possible,
then calculates the length of the other axis