feat: client-side pptx generation (#1)
* build: install pptxgenjs and lodash

* feat: generate pptx on client side and not server

This commit now does the pptx generation on the client side instead of the server.
This removes the need for the server. yay!

* fix: remove cropped image successfully

* fix(crop): close when close button is clicked

Didn't implement closing functionality for cropping

* chore: change favicon and update html title

* remove server folder

* feat: require incident number field

* fix: preserve aspect ratio when adding image to slide

Previously, the scaling formula was broken. So, I used the provided
`sizing` property in the `addImage` method to do the scaling for me.

Problem now is that the photo will have unnecessary empty spaces in it,
which would show up if it was saved

* fix(cropper): decrease modal width

On mobile, the modal width is so wide that when the user tries to
adjust the cropping boundary, they'd end up touching the ends that
interacts with the browser's navigation.

This commit decreases the width such that when the user adjusts
the crop they'd be doing just that

* feat(pptx): improve image aspect ratio maintaining function

Previously, the way the aspect ratio was maintained introduced unecessary
padding to the image which would show up as empty space if it was saved as an image.

So to fix that the image is now scaled manually instead of leaving it to
`pptxgenjs` library. It compares the scale value of both axis, and then
if one of the scale is smaller, it then sets that respective axis to fill up as much as possible,
then calculates the length of the other axis

* housekeeping: remove commented code

* feat: change call text to others