(WIP) feat: create multiple incidents
This commit now allows users to create multiple incidents.

However some drawbacks (for now) are that the initially created incidents
aren't persistent, so they'd be wiped the moment the user relaods the page.

Another drawback is that they can't edit existing incidents.

Planning to use IndexedDB to solve the persistence issue