fix: remove blank photo previews on incident cards
Photo previews in the `<IncidentCard />` components sometimes show
blanks. This was because of the way we implemented how we wanted to show
the photos.

What was initially done was the `<IncidentCard />` component takes in
an `incident` object, which has properties other than the photo keys (i.e
detector, sub_alarm_panel, etc.). We'd take out the `incident_no` property,
leaving only the photo category properties, and then we'd enumerate and display
the photo previews.

Problem is that when an `incident` object has more properties (i.e dispatch time,
arrival time, etc.). The properties weren't accounted for and the code wasn't updated,
hence the blank cards.

The proper fix would be to pick the photo category properties insted and then enumerate
through them